split test by type, update workflow definitions
