+++ .github/workflows/ci.yml
@@ -80,7 +80,6 @@ jobs:
       QISKIT_IBM_PRIVATE_HGP: ${{ secrets.QISKIT_IBM_PRIVATE_HGP }}
       LOG_LEVEL: DEBUG
       STREAM_LOG: True
-      QISKIT_TESTS: skip_online
       QISKIT_IN_PARALLEL: True
     steps:
       - uses: actions/checkout@v2
@@ -96,7 +95,7 @@ jobs:
           pip install -c constraints.txt -e .
           pip install -U -c constraints.txt -r requirements-dev.txt
       - name: Run unit tests
-        run: make test-coverage
+        run: make unit-test-coverage
       - name: Report coverage to coveralls.io
         uses: coverallsapp/github-action@1.1.3
         with:
@@ -136,7 +135,7 @@ jobs:
           pip install -c constraints.txt -e .
           pip install -U -c constraints.txt -r requirements-dev.txt
       - name: Run Tests
-        run: make test
+        run: make integration-test
   tests-finished:
     name: Submit code coverage metrics
     needs: [ unit-tests ]
