--- .github/workflows/cron-staging.yml
@@ -1,41 +1,0 @@
-# This code is part of Qiskit.
-#
-# (C) Copyright IBM 2021.
-#
-# This code is licensed under the Apache License, Version 2.0. You may
-# obtain a copy of this license in the LICENSE.txt file in the root directory
-# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
-#
-# Any modifications or derivative works of this code must retain this
-# copyright notice, and modified files need to carry a notice indicating
-# that they have been altered from the originals.
-
-name: Cron-staging
-on:
-  schedule:
-    - cron: '0 4 * * *'
-jobs:
-  runtime-integration:
-    name: runtime-integration
-    runs-on: macOS-latest
-    env:
-      QISKIT_IBM_STAGING_API_TOKEN: ${{ secrets.QISKIT_IBM_STAGING_API_TOKEN }}
-      QISKIT_IBM_STAGING_API_URL: ${{ secrets.QISKIT_IBM_STAGING_API_URL }}
-      QISKIT_IBM_STAGING_RUNTIME_DEVICE: ${{ secrets.QISKIT_IBM_STAGING_RUNTIME_DEVICE }}
-      QISKIT_IBM_USE_STAGING_CREDENTIALS: True
-      LOG_LEVEL: DEBUG
-      STREAM_LOG: True
-      QISKIT_IN_PARALLEL: True
-    steps:
-      - uses: actions/checkout@v2
-      - name: Set up Python 3.8
-        uses: actions/setup-python@v2
-        with:
-          python-version: 3.8
-      - name: Install Deps
-        run: |
-          python -m pip install --upgrade pip
-          pip install -c constraints.txt -e .
-          pip install -U -c constraints.txt -r requirements-dev.txt
-      - name: Run Tests
-        run: make runtime_integration
