+++ test/integration/test_ibm_provider.py
@@ -73,37 +73,6 @@ class TestIBMProviderEnableAccount(IBMTestCase):
             IBMProvider(qe_token, qe_url, proxies=proxies)
         self.assertIn("ProxyError", str(context_manager.exception))
 
-    def test_provider_init_non_auth_url(self):
-        """Test initializing IBMProvider with a non-auth URL."""
-        qe_token = "invalid"
-        qe_url = API_URL
-
-        with self.assertRaises(IBMProviderCredentialsInvalidUrl) as context_manager:
-            IBMProvider(token=qe_token, url=qe_url)
-
-        self.assertIn("authentication URL", str(context_manager.exception))
-
-    def test_provider_init_non_auth_url_with_hub(self):
-        """Test initializing IBMProvider with a non-auth URL containing h/g/p."""
-        qe_token = "invalid"
-        qe_url = API_URL + "/Hubs/X/Groups/Y/Projects/Z"
-
-        with self.assertRaises(IBMProviderCredentialsInvalidUrl) as context_manager:
-            IBMProvider(token=qe_token, url=qe_url)
-
-        self.assertIn("authentication URL", str(context_manager.exception))
-
-    def test_provider_init_no_credentials(self):
-        """Test initializing IBMProvider with no credentials."""
-        with custom_qiskitrc(), self.assertRaises(
-            IBMProviderCredentialsNotFound
-        ) as context_manager, no_envs(CREDENTIAL_ENV_VARS):
-            IBMProvider()
-
-        self.assertIn(
-            "No IBM Quantum credentials found.", str(context_manager.exception)
-        )
-
     @requires_qe_access
     def test_discover_backend_failed(self, qe_token, qe_url):
         """Test discovering backends failed."""
@@ -129,15 +98,6 @@ class TestIBMProviderAccounts(IBMTestCase):
         super().setUpClass()
         cls.token = "API_TOKEN"
 
-    def test_save_account(self):
-        """Test saving an account."""
-        with custom_qiskitrc():
-            IBMProvider.save_account(self.token, url=QISKIT_IBM_API_URL)
-            stored_cred = IBMProvider.saved_account()
-
-        self.assertEqual(stored_cred["token"], self.token)
-        self.assertEqual(stored_cred["url"], QISKIT_IBM_API_URL)
-
     @requires_qe_access
     def test_provider_init_saved_account(self, qe_token, qe_url):
         """Test initializing IBMProvider with credentials from qiskitrc file."""
@@ -152,61 +112,6 @@ class TestIBMProviderAccounts(IBMTestCase):
         self.assertIsInstance(provider, IBMProvider)
         self.assertEqual(provider.backend._default_hgp.credentials.token, qe_token)
         self.assertEqual(provider.backend._default_hgp.credentials.auth_url, qe_url)
-
-    def test_save_account_specified_provider(self):
-        """Test saving an account with a specified hub/group/project."""
-        default_hgp_to_save = "ibm-q/open/main"
-
-        with custom_qiskitrc() as custom_qiskitrc_cm:
-            hgp_id = HubGroupProjectID.from_stored_format(default_hgp_to_save)
-            IBMProvider.save_account(
-                token=self.token,
-                url=QISKIT_IBM_API_URL,
-                hub=hgp_id.hub,
-                group=hgp_id.group,
-                project=hgp_id.project,
-            )
-
-            # Ensure the `default_provider` name was written to the config file.
-            config_parser = ConfigParser()
-            config_parser.read(custom_qiskitrc_cm.tmp_file.name)
-
-            for name in config_parser.sections():
-                single_credentials = dict(config_parser.items(name))
-                self.assertIn("default_provider", single_credentials)
-                self.assertEqual(
-                    single_credentials["default_provider"], default_hgp_to_save
-                )
-
-    def test_save_account_specified_provider_invalid(self):
-        """Test saving an account without specifying all the hub/group/project fields."""
-        invalid_hgp_ids_to_save = [
-            HubGroupProjectID("", "default_group", ""),
-            HubGroupProjectID("default_hub", None, "default_project"),
-        ]
-        for invalid_hgp_id in invalid_hgp_ids_to_save:
-            with self.subTest(invalid_hgp_id=invalid_hgp_id), custom_qiskitrc():
-                with self.assertRaises(IBMProviderValueError) as context_manager:
-                    IBMProvider.save_account(
-                        token=self.token,
-                        url=QISKIT_IBM_API_URL,
-                        hub=invalid_hgp_id.hub,
-                        group=invalid_hgp_id.group,
-                        project=invalid_hgp_id.project,
-                    )
-                self.assertIn(
-                    "The hub, group, and project parameters must all be specified",
-                    str(context_manager.exception),
-                )
-
-    def test_delete_account(self):
-        """Test deleting an account."""
-        with custom_qiskitrc():
-            IBMProvider.save_account(self.token, url=QISKIT_IBM_API_URL)
-            IBMProvider.delete_account()
-            stored_cred = IBMProvider.saved_account()
-
-        self.assertEqual(len(stored_cred), 0)
 
     @requires_qe_access
     def test_load_account_saved_provider(self, qe_token, qe_url):
@@ -283,30 +188,6 @@ class TestIBMProviderAccounts(IBMTestCase):
                 str(context_manager.exception),
             )
 
-    def test_load_saved_account_invalid_hgp_format(self):
-        """Test loading an account that contains a saved provider in an invalid format."""
-        # Format {'test_case_input': 'error message from raised exception'}
-        invalid_hgps = {
-            "hub_group_project": 'Use the "<hub_name>/<group_name>/<project_name>" format',
-            "default_hub//default_project": "Every field must be specified",
-            "default_hub/default_group/": "Every field must be specified",
-        }
-
-        for invalid_hgp, error_message in invalid_hgps.items():
-            with self.subTest(invalid_hgp=invalid_hgp):
-                with custom_qiskitrc() as temp_qiskitrc, no_envs(CREDENTIAL_ENV_VARS):
-                    # Save the account.
-                    IBMProvider.save_account(token=self.token, url=QISKIT_IBM_API_URL)
-                    # Add an invalid provider field to the account stored.
-                    with open(
-                        temp_qiskitrc.tmp_file.name, "a", encoding="utf-8"
-                    ) as _file:
-                        _file.write("default_provider = {}".format(invalid_hgp))
-                    # Ensure an error is raised if the stored provider is in an invalid format.
-                    with self.assertRaises(IBMProviderError) as context_manager:
-                        IBMProvider()
-                    self.assertIn(error_message, str(context_manager.exception))
-
     @requires_qe_access
     def test_active_account(self, qe_token, qe_url):
         """Test get active account"""
@@ -315,21 +196,6 @@ class TestIBMProviderAccounts(IBMTestCase):
         self.assertIsNotNone(active_account)
         self.assertEqual(active_account["token"], qe_token)
         self.assertEqual(active_account["url"], qe_url)
-
-    def test_save_token_invalid(self):
-        """Test saving an account with invalid tokens. See #391."""
-        invalid_tokens = [None, "", 0]
-        for invalid_token in invalid_tokens:
-            with self.subTest(invalid_token=invalid_token):
-                with self.assertRaises(
-                    IBMProviderCredentialsInvalidToken
-                ) as context_manager:
-                    IBMProvider.save_account(
-                        token=invalid_token, url=QISKIT_IBM_API_URL
-                    )
-                self.assertIn(
-                    "Invalid IBM Quantum token found", str(context_manager.exception)
-                )
 
 
 class TestIBMProviderHubGroupProject(IBMTestCase):
