+++ test/unit/test_registration.py
